arrumando algumas migrations que quebraram na hora de subir pra hostinger

+++ back/database/migrations/2024_05_20_211407_create_rifa_numbers_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('rifa_numbers', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('rifa_id');
+            $table->unsignedBigInteger('rifa_id');
             $table->foreign("rifa_id")->references("id")->on("rifas")->onDelete("cascade");
-            $table->unsignedInteger('client_id');
+            $table->unsignedBigInteger('client_id');
             $table->foreign("client_id")->references("id")->on("clients")->onDelete("cascade");
             $table->json('numbers');
             $table->timestamps();
