+++ back/database/migrations/2024_05_21_174800_add_column_cota_id_to_rifa_numbers_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('rifa_numbers', function (Blueprint $table) {
-            $table->unsignedInteger('cota_id');
+            $table->unsignedBigInteger('cota_id');
             $table->foreign("cota_id")->references("id")->on("cotas")->onDelete("cascade");
         });
     }
